+++ tango/integrations/pytorch_lightning/train.py
@@ -82,6 +82,7 @@ class LightningTrainStep(Step):
     DETERMINISTIC: bool = True
     CACHEABLE = True
     FORMAT: Format = TorchFormat()
+    RESULT_KIND = "model"
 
     def run(  # type: ignore[override]
         self,